split command group to single file
